refactor/reorganise step components
