+++ client/components/recycleAndReuseComponents/Steps/index.js
@@ -1,0 +1,4 @@
+export { AddItem } from "./StepOne/AddItem";
+export { VerifyItem } from "./StepTwo/VerifyItem";
+export { Location } from "./StepThree/Location";
+export { TakeAction } from "./StepThree/TakeAction";
